Refactor schemas to new file
