+++ package.json
@@ -4,6 +4,7 @@
   "description": "yeah you know me",
   "main": "index.js",
   "scripts": {
+    "start": "nodemon ./server/server.js",
     "test": "jest"
   },
   "repository": {
@@ -22,6 +23,7 @@
   },
   "dependencies": {
     "express": "^4.17.1",
-    "mongoose": "^5.13.2"
+    "mongoose": "^5.13.2",
+    "nodemon": "^2.0.12"
   }
 }
